initial overview to set up leveraging dynamo db for personal photos
