+++ db.js
@@ -1,5 +1,6 @@
-import AWS from "aws-sdk";
+const AWS = require("aws-sdk");
 
+const PHOTO_DYNAMO_TABLE = "vwbthree--photos--test";
 AWS.config.update({
     accessKeyId: process.env.AWS_PHOTOS_DB_ACCESS_KEY, // Do NOT HARD-CODE your secret credentials here
     secretAccessKey: process.env.AWS_PHOTOS_DB_SECRET_KEY, // Do NOT HARD-CODE your secret credentials here
@@ -9,4 +10,4 @@ AWS.config.update({
 // Create DynamoDB service object
 const db = new AWS.DynamoDB.DocumentClient({ apiVersion: "latest" });
 
-export default db;
+module.exports = { db, PHOTO_DYNAMO_TABLE };
