+++ package.json
@@ -30,6 +30,7 @@
         "throttle-debounce": "^2.1.0"
     },
     "devDependencies": {
-        "@squoosh/lib": "^0.4.0"
+        "@squoosh/lib": "^0.4.0",
+        "csvtojson": "^2.0.10"
     }
 }
