+++ pages/index.js
@@ -1,8 +1,8 @@
-import React, { useState, useEffect, useLayoutEffect} from "react";
+import React, { useState, useEffect, useLayoutEffect } from "react";
 
 import Link from "next/link";
 import Layout from "../components/Layout";
-import FadeIn from '../components/FadeIn';
+import FadeIn from "../components/FadeIn";
 
 const LandingBackground = props => {
     const [{ width, height }, setDimensions] = useState({
@@ -55,8 +55,8 @@ const Menu = props => {
             style={{
                 width: "100%",
                 height: `${height}px`,
-                display: 'flex',
-                alignItems: 'center'
+                display: "flex",
+                alignItems: "center"
             }}
         >
             {props.children}
@@ -101,8 +101,8 @@ const Home = () => {
     useEffect(() => {
         setTimeout(() => {
             setMenuVisible(true);
-        }, 500)
-    })
+        }, 500);
+    });
 
     return (
         <div
@@ -115,29 +115,30 @@ const Home = () => {
                 textColor="text-white"
             >
                 <LandingBackground />
-                    <FadeIn>
-                        <Menu isVisible={isMenuVisible}>
-                            <section className="flex items-center flex-col w-full">
-                                <div
-                                    className="text-center text-gray-200 font-light flex flex-col items-center"
-                                    style={{
-                                        height: "200px",
-                                        width: "200px",
-                                        fontFamily: "Simsun"
-                                    }}
-                                >
-                                    <Link href="/photos">
-                                        <LinkButton passHref>Photography</LinkButton>
-                                    </Link>
-                                    <div className="fade-in border border-solid border-gray-200 w-6" />
-                                    <Link href="/about">
-                                        <LinkButton passHref>About</LinkButton>
-                                    </Link>
-                                </div>
-                            </section>
-                        </Menu>
-                    </FadeIn>
-
+                <FadeIn>
+                    <Menu isVisible={isMenuVisible}>
+                        <section className="flex items-center flex-col w-full">
+                            <div
+                                className="text-center text-gray-200 font-light flex flex-col items-center"
+                                style={{
+                                    height: "200px",
+                                    width: "200px",
+                                    fontFamily: "Simsun"
+                                }}
+                            >
+                                <Link href="/photos/collections">
+                                    <LinkButton passHref>
+                                        Photography
+                                    </LinkButton>
+                                </Link>
+                                <div className="fade-in border border-solid border-gray-200 w-6" />
+                                <Link href="/about">
+                                    <LinkButton passHref>About</LinkButton>
+                                </Link>
+                            </div>
+                        </section>
+                    </Menu>
+                </FadeIn>
             </Layout>
         </div>
     );
