+++ pages/photos/[id]/index.js
@@ -1,159 +1,228 @@
-import React from "react";
-import Link from "next/link";
-
-import { FixedSizeList as List } from "react-window";
-import AutoSizer from "react-virtualized-auto-sizer";
+import { useEffect, useContext, useState } from "react";
+import { useRouter } from "next/router";
+import { GiCrane } from "react-icons/gi";
+import { AiOutlinePlus, AiOutlineMinus } from "react-icons/ai";
 
 import Layout from "../../../components/Layout";
 import PhotoView from "../../../components/Photo";
+import Panel, { PanelContext } from "../../../components/Panel";
+import Divider from "../../../components/Divider";
+import { Expand } from "../../../components/Expand";
 
-import { generatePhotoSlug } from "../../../lib/photos";
+import { db, PHOTO_DYNAMO_TABLE } from "../../../db";
 
-import {
-    getAllCollectionPaths,
-    getCollectionData,
-    getCollection
-} from "../../../lib/collections";
+import { getAllPhotoPaths, getPhotoData } from "../../../lib/photos";
+import { getCollection } from "../../../lib/collections";
+import DownArrow from "../../../components/DownArrrow";
 
-const ScrollToNext = ({ onClick }) => (
-    <div
-        className="absolute w-full text-center text-gray-900 z-20"
-        style={{ bottom: "5%" }}
-    >
-        <button className="focus:outline-none" onClick={onClick}>
-            ↓
-        </button>
+const DetailHeader = props => (
+    <div className="mx-4">
+        <div
+            style={{
+                fontFamily: "Gill Sans"
+            }}
+        >
+            {props.children}
+        </div>
     </div>
 );
 
-const storeIndex = (index, collectionName) => {
-    localStorage.setItem(
-        "vwb_photos_index",
-        JSON.stringify({ collectionName, index })
-    );
-};
+const Background = props => {
+    const router = useRouter();
 
-const PhotoContainer = ({
-    photo,
-    index,
-    collection,
-    isLast,
-    handleNextPhotoClick
-}) => {
+    const handleClickBack = () => {
+        if (document.referrer.indexOf(`vwbthree`) > -1) {
+            router.back();
+        }
+    };
+
     return (
-        <div className="h-full w-full relative flex items-center content-center overflow-hidden">
-            <PhotoView
-                photo={photo}
-                isRaised={true}
-                style={{
-                    height: "60%"
-                }}
-                render={image => (
-                    <Link
-                        href="/photos/[id]/[slug]"
-                        as={`/photos/${collection.slug}/${generatePhotoSlug(
-                            photo
-                        )}`}
-                    >
-                        <a onClick={() => storeIndex(index, collection.name)}>
-                            {image}
-                        </a>
-                    </Link>
-                )}
-            />
-            {!isLast && (
-                <ScrollToNext onClick={() => handleNextPhotoClick(index)} />
-            )}
+        <div className="w-full h-screen" onClick={() => handleClickBack()}>
+            <div style={{ height: "100%" }} className="w-full fixed">
+                <div className="flex items-center overflow-hidden relative h-full">
+                    {props.children}
+                </div>
+            </div>
         </div>
     );
 };
 
-const PhotoGroupPage = props => {
-    const [windowInnerHeight, setInnerHeight] = React.useState(null);
-    const [hasScrolled, setHasScrolled] = React.useState(false);
-    const ref = React.useRef();
-
-    const setRef = element => {
-        ref.current = element;
-    };
-
-    React.useEffect(() => {
-        setInnerHeight(document.documentElement.clientHeight);
+const FadeIn = props => {
+    const [isMounted, setMounted] = useState(false);
+    useEffect(() => {
+        setMounted(true);
     }, []);
 
-    React.useEffect(() => {
-        if (windowInnerHeight && ref.current && !hasScrolled) {
-            const potentialIndex = JSON.parse(
-                localStorage.getItem("vwb_photos_index")
-            );
+    return (
+        <div
+            style={{ opacity: `${isMounted ? 1 : 0}`, transition: "all 0.4s" }}
+        >
+            {props.children}
+        </div>
+    );
+};
 
-            if (potentialIndex) {
-                if (potentialIndex?.collectionName === props.collection.name) {
-                    ref?.current?.scrollTo(
-                        windowInnerHeight * potentialIndex.index
-                    );
-                }
+const ContactForm = props => {
+    const [isSubmitted, setSubmitted] = useState(false);
 
-                if (ref.current) {
-                    localStorage.removeItem("vwb_photos_index");
-                }
+    return (
+        <section className="pt-4 px-4 w-full sm:w-3/4 lg:w-2/3 xl:w-1/2 mx-auto">
+            {!isSubmitted && (
+                <section className="flex flex-col items-center text-center">
+                    <section className="flex flex-col items-center pt-12">
+                        <GiCrane size={90} />
+                        <p className="text-lg pt-2 pb-6">
+                            Shop under construction
+                        </p>
+                        <Expand
+                            renderHeader={(isExpanded, setExpanded) => {
+                                return (
+                                    <button
+                                        className=" px-2 py-1 text-gray-900 mt-2 flex flex-row items-center"
+                                        onClick={() => setExpanded(!isExpanded)}
+                                    >
+                                        <p className="text-sm pr-4">
+                                            Still want one?
+                                        </p>
+                                        <section className="">
+                                            {isExpanded ? (
+                                                <AiOutlineMinus size={12} />
+                                            ) : (
+                                                <AiOutlinePlus size={12} />
+                                            )}
+                                        </section>
+                                    </button>
+                                );
+                            }}
+                        >
+                            <p className="text-xs pt-2">
+                                <div>File a custom order!</div>
+                                <div>
+                                    Digital download and custom prints available
+                                    from $50
+                                </div>
+                            </p>
+                        </Expand>
+                    </section>
+                    <section className="flex flex-col pt-8 absolute bottom-0 w-full px-4 max-w-xs">
+                        <p className="text-sm pt-10">Custom orders available</p>
+                        <p></p>
+                        <a
+                            className="my-3 p-3 hover:shadow-xl text-gray-300 focus:outline-none bg-teal-700 rounded"
+                            href={`mailto:vwbthree.photos@gmail.com?subject=Interested%20in:%20${props.photo.title}`}
+                            onClick={() => setSubmitted(true)}
+                        >
+                            Contact
+                        </a>
+                    </section>
+                </section>
+            )}
+            {isSubmitted && (
+                <section className="flex flex-col items-center pt-4">
+                    <p className="text-lg font-semibold text-center">
+                        Thank you for your interest!
+                    </p>
+                    <p className="text-center test-sm pt-2">
+                        You will be contaced shortly
+                    </p>
+                    <button
+                        className="p-3 my-4 border-2 border-solid border-gray-400 rounded"
+                        onClick={() => props.setPanelOpen(false)}
+                    >
+                        Close
+                    </button>
+                </section>
+            )}
+        </section>
+    );
+};
 
-                setHasScrolled(true);
-            }
-        }
-    });
+const PanelContent = props => {
+    const [isPanelOpen, setPanelOpen] = useContext(PanelContext);
 
+    return (
+        <section className="pt-2 h-full">
+            <DetailHeader>
+                <div className="flex items-center justify-between">
+                    <h1 className="text-2xl pl-6">{props.photo.title}</h1>
+                    <div className="flex flex-col items-center">
+                        <button
+                            className={`${
+                                isPanelOpen
+                                    ? "bg-white mt-2 rounded-full border border-gray-300 border-solid p-2 shadow-xl"
+                                    : "bg-teal-700 rounded p-3"
+                            }   text-white focus:outline-none `}
+                            style={{ transition: "all .4s" }}
+                            onClick={() => setPanelOpen(!isPanelOpen)}
+                        >
+                            {isPanelOpen ? (
+                                <DownArrow style={{ height: "25px" }} />
+                            ) : (
+                                "Purchase"
+                            )}
+                        </button>
+                    </div>
+                </div>
+            </DetailHeader>
+            {isPanelOpen && (
+                <div className="pt-4">
+                    <FadeIn>
+                        <Divider />
+                        <ContactForm
+                            setPanelOpen={setPanelOpen}
+                            photo={props.photo}
+                            collection={props.collection}
+                        />
+                    </FadeIn>
+                </div>
+            )}
+        </section>
+    );
+};
+
+const PhotoDetailPage = ({ photo, collection }) => {
     return (
         <Layout
             isOpenDefault={false}
-            navClass="bg-white shadow"
-            title={props.collection.name}
+            navClass="bg-white"
+            title={photo.displayName}
         >
-            <div
-                className={`w-screen ${windowInnerHeight ? "" : "h-screen"}`}
-                style={{ height: windowInnerHeight ? windowInnerHeight : "" }}
-            >
-                <AutoSizer>
-                    {({ height, width }) => (
-                        <List
-                            className="photo-list"
-                            ref={element => setRef(element)}
-                            height={height}
-                            itemCount={props.photos.length}
-                            itemData={props.photos}
-                            itemSize={height}
-                            width={width}
-                        >
-                            {({ index, style }) => {
-                                return (
-                                    <div style={style}>
-                                        <PhotoContainer
-                                            collection={props.collection}
-                                            photo={props.photos[index]}
-                                            index={index}
-                                            isLast={
-                                                index ===
-                                                props.photos.length - 1
-                                            }
-                                            handleNextPhotoClick={currentIndex => {
-                                                ref.current.scrollToItem(
-                                                    currentIndex + 1
-                                                );
-                                            }}
-                                        />
-                                    </div>
-                                );
+            <Background>
+                <PhotoView
+                    photo={photo}
+                    style={{
+                        height: "80%"
+                    }}
+                    render={image => (
+                        <div
+                            onClick={e => {
+                                e.stopPropagation();
                             }}
-                        </List>
+                        >
+                            {image}
+                        </div>
                     )}
-                </AutoSizer>
-            </div>
+                />
+            </Background>
+            <Panel>
+                <PanelContent photo={photo} />
+            </Panel>
         </Layout>
     );
 };
 
 export async function getStaticPaths() {
-    const paths = getAllCollectionPaths();
+    var params = {
+        TableName: PHOTO_DYNAMO_TABLE
+    };
+
+    var result = await db.scan(params).promise();
+    const photos = result.Items;
+
+    const paths = photos.map(item => ({
+        params: { id: `${item.displayName}--${item.photoName}` }
+    }));
+
     return {
         paths,
         fallback: false
@@ -161,9 +230,29 @@ export async function getStaticPaths() {
 }
 
 export async function getStaticProps({ params }) {
-    const photos = getCollectionData(params.id);
-    const collection = getCollection(params.id);
-    return { props: { photos, collection } };
+    const id = params.id.split("--")[1];
+
+    var params = {
+        TableName: PHOTO_DYNAMO_TABLE,
+        ExpressionAttributeValues: {
+            ":photoName": id
+        },
+        KeyConditionExpression: "photoName = :photoName"
+    };
+
+    try {
+        const result = await db.query(params).promise();
+        const photo = result.Items[0];
+
+        console.log(photo);
+
+        return { props: { photo } };
+    } catch (e) {
+        console.error("failed");
+        console.error(e);
+    }
+
+    return { props: { photo: [] } };
 }
 
-export default PhotoGroupPage;
+export default PhotoDetailPage;
